+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Tranform migration function: Add the `onlyInitializer` modifier to any function which name ends with `_migrate`.
+
 ## 0.4.1 (2025-09-18)
 
 - Revert part 0.4.0: Do not automatically delete initialisable contracts when "delete originals" is enabled (in peer project mode).